added snapshot listener in interactor, created items from db output

+++ app/src/main/java/com/example/michl/aimission/Models/AimItem.kt
@@ -4,7 +4,7 @@ data class AimItem(
         var id:String,     // this should be the id which firebase uses to save this item in table aims
         var title:String,
         var description:String,
-        var repeatCount:Int, // if aim is marked as a repeated aim, save here the amount of numbers this aim should be absolved
+        var repeatCount:Long, // if aim is marked as a repeated aim, save here the amount of numbers this aim should be absolved
         var isHighPriority:Boolean,
         var status: Status,
         var genre:Genre
@@ -14,6 +14,7 @@ data class AimItem(
 enum class Status
 {
     OPEN,PROGRESS,DONE
+
 }
 
 enum class Genre
